Unbound wildcard range query cripples es
A couple weeks back when experimenting with the metrics api we suddenly had 40 ES nodes completely pegged on Mem/CPU/IO (and swapping). This lasted for around 6-8 hours before the cluster eventually returned to a healthy state. My past experience with lucene made me suspect this query may have been the cause but was unable to determine at the time if it was the issue. Ran again today a single time and it pegged our cluster for about 30 minutes before returning to a healthy state. This may only apply to larger installations of ES (we have 50+ billion docs) but it's likely better to use the stats API regardless.